Move tabulate and six to separate folders: Just to make things more cleaner and adds license for six
