+++ pgmpy/extern/six/__init__.py
@@ -1,0 +1,3 @@
+import six
+
+__all__ = ['six']
