+++ pgmpy/extern/tabulate/__init__.py
@@ -1,0 +1,3 @@
+from .tabulate import tabulate
+
+__all__ = ['tabulate',]
